Reorganize send_vars_to_gpu and move setup_pagefault_handler to be after queue declaration
